+++ lib/Utils/index.d.ts
@@ -56,5 +56,7 @@ export * from "./qr-render.js";
 export * from "./auto-reconnect.js";
 export * from "./debug-info.js";
 export * from "./ffmpeg-path.js";
+export * from "./doctor.js";
+export { printBanner } from "./banner.js";
 
 export * from "./MessageBuilder.js";